+++ CS2MapView.Exporter/Setting.cs
@@ -9,7 +9,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using UnityEngine.Scripting;
-using CS2MapView.Exporter.System;
+using CS2MapView.Exporter.Systems;
 
 namespace CS2MapView.Exporter
 {
